refactor: 상태관리와 검색 및 필터링 조건 url 동기화를 context api에서 커스텀훅으로 변경
